Spec namespaces
